requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.28.0

# MijiaAPI dependencies
mijiaAPI>=2.0.1
aiohttp>=3.8.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0